+++ gdextension/src/mresource.h
@@ -219,7 +219,7 @@ class MResource : public Resource {
     void decompress_qtq_rf(const PackedByteArray& compress_data,PackedByteArray& uncompress_data,uint32_t window_width,uint32_t decompress_index);
     // Will add and remove Linear Regression with least square method
     Vector<uint32_t> flatten_ols(float* data,uint32_t witdth,uint16_t devision); // Will ignore holes
-    void unflatten_ols(float* data,uint32_t witdth,uint8_t devision,const Vector<uint32_t>& headers); 
+    void unflatten_ols(float* data,uint32_t witdth,uint16_t devision,const Vector<uint32_t>& headers); 
     uint32_t flatten_section_ols(float* data,MPixelRegion px_region,uint32_t window_width,Basis matrix_a_invers);
     void unflatten_section_ols(float* data,MPixelRegion px_region,uint32_t window_width,uint32_t header);
     _FORCE_INLINE_ uint64_t get_sumx(uint64_t n);
